added a post route for listings
